LayoutTests/fast/frames/seamless/seamless-form-get-override.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner)
    testRunner.waitUntilDone();

window.addEventListener('message', function(evt) {
    shouldBeEqualToString("frames[0].frames[0].document.body.textContent", "DONE\n");

    if (window.testRunner)
        testRunner.notifyDone();
}, false);
</script>
<iframe srcdoc="
    Waiting...
    <script>
        window.onload=function() {
            frames[0].document.getElementsByTagName('form')[0].submit();
        }
    </script>
    <iframe seamless
            srcdoc='<form target=_self method=GET action=resources/done.html>
                        <input type=submit>
                    </from>
    '></iframe>
"></iframe>
